feat: add auth session and flows contexts
