+++ components/AppContainer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { Alert, SafeAreaView, StatusBar, StyleSheet } from 'react-native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { useTheme } from '@/contexts/ThemeContext';
-import { useAuth } from '@/contexts/AuthContext';
+import { useAuthFlows } from '@/contexts/AuthFlowsContext';
 import { useTranslation } from '@/contexts/I18nContext';
 import usePushNotifications from '@/hooks/usePushNotifications';
 import useDailyQuote from '@/hooks/useDailyQuote';
@@ -11,7 +11,7 @@ export const AppContainer: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
   const { theme } = useTheme();
-  const { authError, setAuthError } = useAuth();
+  const { authError, setAuthError } = useAuthFlows();
   const { t } = useTranslation();
   const backgroundColor = theme.background;
   const barStyle =
